refactor: added getFile util method for tests
