+++ src/plugin.js
@@ -14,12 +14,9 @@ class ESBuildPlugin {
     }
 
     compiler.hooks.thisCompilation.tap('esbuild', (compilation) => {
-      compilation.hooks.childCompiler.tap(
-        'esbuild',
-        (childCompiler) => {
-          childCompiler.$esbuildService = compiler.$esbuildService
-        }
-      )
+      compilation.hooks.childCompiler.tap('esbuild', (childCompiler) => {
+        childCompiler.$esbuildService = compiler.$esbuildService
+      })
     })
 
     compiler.hooks.run.tapPromise('esbuild', async () => {
